add conftest.py file to share synthetic data creation across test files
